+++ src/config/gro.config.default.ts
@@ -4,6 +4,7 @@ import type {GroConfigCreator, PartialGroConfig} from './config.js';
 import {LogLevel} from '../utils/log.js';
 import {PartialBuildConfig} from './buildConfig.js';
 import {pathExists} from '../fs/nodeFs.js';
+import {basePathToSourceId, toBuildExtension} from '../paths.js';
 
 // This is the default config that's used if the current project does not define one.
 // The default config detects
@@ -11,6 +12,10 @@ import {pathExists} from '../fs/nodeFs.js';
 // if it sees both a `src/index.html` and `src/index.ts`.
 // It also looks for a primary Node server entry point at `src/server/server.ts`.
 // Both are no-ops if not detected.
+
+export const SERVER_SOURCE_BASE_PATH = 'server/server.ts';
+export const SERVER_BUILD_BASE_PATH = toBuildExtension(SERVER_SOURCE_BASE_PATH); // 'server/server.js'
+export const SERVER_SOURCE_ID = basePathToSourceId(SERVER_SOURCE_BASE_PATH); // '/home/to/your/src/server/server.ts'
 
 const createConfig: GroConfigCreator = async () => {
 	const config: PartialGroConfig = {
@@ -20,7 +25,7 @@ const createConfig: GroConfigCreator = async () => {
 				name: 'node',
 				platform: 'node',
 				input: [
-					(await pathExists('src/server/server.ts')) ? 'server/server.ts' : null!,
+					(await hasGroServer()) ? SERVER_SOURCE_BASE_PATH : null!,
 					createFilter('**/*.{task,test,config,gen}*.ts'),
 				].filter(Boolean),
 			},
@@ -41,11 +46,14 @@ const toDefaultBrowserBuild = (): PartialBuildConfig => ({
 	dist: true,
 });
 
-// TODO extract helper?
-const hasDeprecatedGroFrontend = async (): Promise<boolean> => {
+// TODO extract helper? or is this default config file its actual home?
+export const hasDeprecatedGroFrontend = async (): Promise<boolean> => {
 	const [hasIndexHtml, hasIndexTs] = await Promise.all([
 		pathExists('src/index.html'),
 		pathExists('src/index.ts'),
 	]);
 	return hasIndexHtml && hasIndexTs;
 };
+
+// TODO extract helper? or is this default config file its actual home?
+export const hasGroServer = (): Promise<boolean> => pathExists(SERVER_SOURCE_ID);